requirements.txt
numpy
matplotlib
cython
tensorflow-gpu
keras
keyboard
Pillow
pandas
opencv-python
